Modular UI Strategy panel
