+++ key.core/src/main/java/de/uka/ilkd/key/strategy/JavaCardDLStrategyFactory.java
@@ -29,13 +29,6 @@ public class JavaCardDLStrategyFactory implements StrategyFactory {
      */
     public static final Name NAME = new Name(JavaCardDLStrategy.JAVA_CARD_DL_STRATEGY);
 
-    public static final String TOOL_TIP_STOP_AT_DEFAULT =
-        "<html>Stop when (i) the maximum number of rule<br>"
-            + "applications is reached or (ii) no more rules are<br>"
-            + "applicable on the proof tree.</html>";
-    public static final String TOOL_TIP_STOP_AT_UNCLOSABLE =
-        "<html>Stop as soon as the first not automatically<br>"
-            + "closable goal is encountered.</html>";
     public static final String TOOL_TIP_OSS_ON =
         "<html>" + "Turns on One Step Simplification. This will result in<br>"
             + "(sometimes significantly) shorter proofs which,<br>"
@@ -189,26 +182,6 @@ public class JavaCardDLStrategyFactory implements StrategyFactory {
     public static final String TOOL_TIP_EXPAND_LOCAL_QUERIES_OFF =
         "<html>" + "Expansion of local queries is turned off. <br>"
             + "This setting is independent of the query treatment setting." + "</html>";
-    public static final String TOOL_TIP_ARITHMETIC_BASE = "<html>" + "Basic arithmetic support:"
-        + "<ul>" + "<li>Simplification of polynomial expressions</li>"
-        + "<li>Computation of Gr&ouml;bner Bases for polynomials in the antecedent</li>"
-        + "<li>(Partial) Omega procedure for handling linear inequations</li>" + "</ul>"
-        + "</html>";
-    public static final String TOOL_TIP_ARITHMETIC_DEF_OPS =
-        "<html>" + "Automatically expand defined symbols like:" + "<ul>"
-            + "<li><tt>/</tt>, <tt>%</tt>, <tt>jdiv</tt>, <tt>jmod</tt>, ...</li>"
-            + "<li><tt>int_RANGE</tt>, <tt>short_MIN</tt>, ...</li>"
-            + "<li><tt>inInt</tt>, <tt>inByte</tt>, ...</li>"
-            + "<li><tt>addJint</tt>, <tt>mulJshort</tt>, ...</li>" + "</ul>" + "</html>";
-    public static final String TOOL_TIP_ARITHMETIC_MODEL_SEARCH = "<html>"
-        + "Support for non-linear inequations and model search.<br>" + "In addition, this performs:"
-        + "<ul>" + "<li>Multiplication of inequations with each other</li>"
-        + "<li>Systematic case distinctions (cuts)</li>" + "</ul>"
-        + "This method is guaranteed to find counterexamples for<br>"
-        + "invalid goals that only contain polynomial (in)equations.<br>"
-        + "Such counterexamples turn up as trivially unprovable goals.<br>"
-        + "It is also able to prove many more valid goals involving<br>"
-        + "(in)equations, but will in general not terminate on such goals." + "</html>";
     public static final String TOOL_TIP_QUANTIFIER_NONE =
         "<html>" + "Do not instantiate quantified formulas automatically" + "</html>";
     public static final String TOOL_TIP_QUANTIFIER_NO_SPLITS = "<html>"
@@ -255,29 +228,6 @@ public class JavaCardDLStrategyFactory implements StrategyFactory {
             + "limit the number of inductive proofs." + "</html>";
 
     public JavaCardDLStrategyFactory() {
-    }
-
-    public static final String toolTipUserOff(int i) {
-        return "Taclets of the rule set \"userTaclets" + i + "\" are not applied automatically";
-    }
-
-    public static final String toolTipUserLow(int i) {
-        return "Taclets of the rule set \"userTaclets" + i
-            + "\" are applied automatically with low priority";
-    }
-
-    public static final String toolTipUserHigh(int i) {
-        return "Taclets of the rule set \"userTaclets" + i
-            + "\" are applied automatically with high priority";
-    }
-
-    private static OneOfStrategyPropertyDefinition getStopAt() {
-        return new OneOfStrategyPropertyDefinition(StrategyProperties.STOPMODE_OPTIONS_KEY,
-            "Stop at",
-            new StrategyPropertyValueDefinition(StrategyProperties.STOPMODE_DEFAULT, "Default",
-                TOOL_TIP_STOP_AT_DEFAULT),
-            new StrategyPropertyValueDefinition(StrategyProperties.STOPMODE_NONCLOSE, "Unclosable",
-                TOOL_TIP_STOP_AT_UNCLOSABLE));
     }
 
     private static OneOfStrategyPropertyDefinition getOssUsage() {
@@ -382,17 +332,6 @@ public class JavaCardDLStrategyFactory implements StrategyFactory {
                 TOOL_TIP_QUERY_OFF));
     }
 
-    private static OneOfStrategyPropertyDefinition getArithmeticTreatment() {
-        return new OneOfStrategyPropertyDefinition(StrategyProperties.NON_LIN_ARITH_OPTIONS_KEY,
-            "Arithmetic treatment",
-            new StrategyPropertyValueDefinition(StrategyProperties.NON_LIN_ARITH_NONE, "Basic",
-                TOOL_TIP_ARITHMETIC_BASE),
-            new StrategyPropertyValueDefinition(StrategyProperties.NON_LIN_ARITH_DEF_OPS, "DefOps",
-                TOOL_TIP_ARITHMETIC_DEF_OPS),
-            new StrategyPropertyValueDefinition(StrategyProperties.NON_LIN_ARITH_COMPLETION,
-                "Model Search", TOOL_TIP_ARITHMETIC_MODEL_SEARCH));
-    }
-
     private static OneOfStrategyPropertyDefinition getQuantifierTreatment() {
         return new OneOfStrategyPropertyDefinition(StrategyProperties.QUANTIFIERS_OPTIONS_KEY,
             "Quantifier treatment", 2,
@@ -429,33 +368,9 @@ public class JavaCardDLStrategyFactory implements StrategyFactory {
                 TOOL_TIP_AUTO_INDUCTION_OFF));
     }
 
-    private static OneOfStrategyPropertyDefinition getUserOptions() {
-        // User properties
-        List<AbstractStrategyPropertyDefinition> props = new LinkedList<>();
-        for (int i = 1; i <= StrategyProperties.USER_TACLETS_NUM; ++i) {
-            OneOfStrategyPropertyDefinition user = new OneOfStrategyPropertyDefinition(
-                StrategyProperties.userTacletsOptionsKey(i), i + ":  ",
-                new StrategyPropertyValueDefinition(StrategyProperties.USER_TACLETS_OFF, "Off",
-                    toolTipUserOff(i), 3, 1),
-                new StrategyPropertyValueDefinition(StrategyProperties.USER_TACLETS_LOW,
-                    "Low prior.", toolTipUserLow(i), 4, 2),
-                new StrategyPropertyValueDefinition(StrategyProperties.USER_TACLETS_HIGH,
-                    "High prior.", toolTipUserHigh(i), 6, 2));
-            props.add(user);
-        }
-
-        return new OneOfStrategyPropertyDefinition(null, "User-specific taclet sets",
-            "<html>" + "These options define whether user- and problem-specific taclet sets<br>"
-                + "are applied automatically by the strategy. Problem-specific taclets<br>"
-                + "can be defined in the \\rules-section of a .key-problem file. For<br>"
-                + "automatic application, the taclets have to contain a clause<br>"
-                + "\\heuristics(userTaclets1), \\heuristics(userTaclets2), etc." + "</html>",
-            -1, props.toArray(new AbstractStrategyPropertyDefinition[0]));
-    }
-
     @Override
-    public Strategy<@NonNull Goal> create(Proof proof, StrategyProperties strategyProperties) {
-        return new ModularJavaDLStrategy(proof, strategyProperties);
+    public JavaCardDLStrategy create(Proof proof, StrategyProperties strategyProperties) {
+        return new JavaCardDLStrategy(proof, strategyProperties);
     }
 
     @Override
@@ -466,7 +381,6 @@ public class JavaCardDLStrategyFactory implements StrategyFactory {
     @Override
     public StrategySettingsDefinition getSettingsDefinition() {
         // Properties
-        final OneOfStrategyPropertyDefinition stopAt = getStopAt();
         final OneOfStrategyPropertyDefinition ossUsage = getOssUsage();
         final OneOfStrategyPropertyDefinition proofSplitting = getProofSplitting();
         final OneOfStrategyPropertyDefinition loopTreatment = getLoopTreatment();
@@ -476,15 +390,13 @@ public class JavaCardDLStrategyFactory implements StrategyFactory {
             getMergePointStatementTreatment();
         final OneOfStrategyPropertyDefinition dependencyContracts = getDependencyContracts();
         final OneOfStrategyPropertyDefinition queryTreatment = getQueryTreatment();
-        final OneOfStrategyPropertyDefinition arithmeticTreatment = getArithmeticTreatment();
         final OneOfStrategyPropertyDefinition quantifierTreatment = getQuantifierTreatment();
         final OneOfStrategyPropertyDefinition classAxiom = getClassAxiom();
         final OneOfStrategyPropertyDefinition autoInduction = getAutoInduction();
-        final OneOfStrategyPropertyDefinition userOptions = getUserOptions();
         // Model
-        return new StrategySettingsDefinition("Java DL Options", stopAt, ossUsage, proofSplitting,
+        return new StrategySettingsDefinition("Java DL Options", ossUsage, proofSplitting,
             loopTreatment, blockTreatment, methodTreatment, mergePointStatementTreatment,
-            dependencyContracts, queryTreatment, arithmeticTreatment, quantifierTreatment,
-            classAxiom, autoInduction, userOptions);
+            dependencyContracts, queryTreatment, quantifierTreatment,
+            classAxiom, autoInduction);
     }
 }
